requirements.txt
# Evidence Integrity Validator v1.0
flask>=3.0
Pillow>=10.0
PyPDF2>=3.0
reportlab>=4.0
